example_project/requirements.txt
asgiref==3.5.0; python_version >= "3.8" and python_version < "4.0"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
beautifulsoup4==4.11.1; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.6.0"
bleach==4.1.0; python_version >= "3.8" and python_version < "4.0"
certifi==2021.10.8; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.6.0"
charset-normalizer==2.0.10; python_full_version >= "3.6.0" and python_version >= "3.8" and python_version < "4.0"
django-environ==0.8.1; python_version >= "3.4" and python_version < "4"
django-extensions==3.1.5; python_version >= "3.8" and python_version < "4.0"
django-filter==21.1; python_version >= "3.8" and python_version < "4.0"
django-model-utils==4.2.0
django==4.0.2; python_version >= "3.8"
djangorestframework==3.13.1; python_version >= "3.6"
drf-search-service @ git+https://github.com/digirati-co-uk/drf-search-service.git@feature/GET_search_classes ; python_version >= "3.8" and python_version < "4.0"
idna==3.3; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version >= "3.8" and python_version < "4.0"
ordered-set==4.1.0; python_version >= "3.8" and python_version < "4.0"
packaging==21.3; python_version >= "3.8" and python_version < "4.0"
psycopg2-binary==2.9.3; python_version >= "3.8" and python_version < "4.0"
pyparsing==3.0.7; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.3.0"
pytz==2021.3; python_version >= "3.8" and python_version < "4.0"
requests==2.27.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.3.0"
soupsieve==2.3.2.post1; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.6.0"
sqlparse==0.4.2; python_version >= "3.8" and python_version < "4.0"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8" and python_version < "4.0"
urllib3==1.26.8; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.8"
webencodings==0.5.1; python_version >= "3.8" and python_version < "4.0"
